+++ custom_components/tarifas_energia_brasil/const.py
@@ -13,7 +13,7 @@ from homeassistant.const import Platform
 
 DOMAIN = "tarifas_energia_brasil"
 NAME = "Tarifas Energia Brasil"
-VERSION = "0.1.3"
+VERSION = "0.1.4"
 
 PLATFORMS: list[Platform] = [Platform.SENSOR]